WIP: reordering ffi bindings as they were unordered.
